+++ vcf/src/vcf/FoundationKit/Win32Condition.h
@@ -48,7 +48,7 @@ class Win32Condition : public VCF::Object, public ConditionPeer {
 public:
 
 	enum {
-		MaxQueue = ((unsigned long)-1)/2
+		MaxQueue = ((uint32)-1)/2
 	};
 
 	Win32Condition( Condition* condition );
@@ -75,7 +75,7 @@ protected:
 	HANDLE mutex_;
 
 	unsigned gone_;  // # threads that timed out and never made it to queue_
-    unsigned long blocked_; // # threads blocked on the condition
+    uint32 blocked_; // # threads blocked on the condition
     unsigned waiting_; // # threads no longer waiting for the condition but
                         // still waiting to be removed from queue_
 };
